Extract postgres.js connection and query-span helpers into postgres-js-utils
